fix: 使用 AutoMirrored 图标，修复 CharacterPresetInitializer 冗余转换

+++ app/app/src/main/java/com/echonion/nion/preset/CharacterPresetInitializer.kt
@@ -167,7 +167,7 @@ object CharacterPresetInitializer {
 
                 inputStream.use { input ->
                     FileOutputStream(destFile).use { output ->
-                        fileSize = input.copyTo(output).toLong()
+                        fileSize = input.copyTo(output)
                     }
                 }
